Float type alpha

+++ src/PixelDissimilarity.cpp
@@ -1,8 +1,8 @@
 #include <PixelDissimilarity.hpp>
 
-template <class Pixel> double PixelDissimilarity<Pixel>::L1(ImgIdx index1, ImgIdx index2) const {
+template <class Pixel> float PixelDissimilarity<Pixel>::L1(ImgIdx index1, ImgIdx index2) const {
     if (_channels == 1)
-        return std::abs((double)_image[index1] - (double)_image[index2]);
+        return (float)std::abs((double)_image[index1] - (double)_image[index2]);
 
     double dist = 0.0;
     for (ImgIdx ch = 0; ch < _channels; ch++) {
@@ -10,12 +10,12 @@ template <class Pixel> double PixelDissimilarity<Pixel>::L1(ImgIdx index1, ImgId
         double diff = std::abs((double)_image[index1 + offset] - (double)_image[index2 + offset]);
         dist += diff;
     }
-    return dist / (double)_channels;
+    return (float)(dist / (double)_channels);
 }
 
-template <class Pixel> double PixelDissimilarity<Pixel>::L2(ImgIdx index1, ImgIdx index2) const {
+template <class Pixel> float PixelDissimilarity<Pixel>::L2(ImgIdx index1, ImgIdx index2) const {
     if (_channels == 1)
-        return std::abs((double)_image[index1] - (double)_image[index2]);
+        return (float)std::abs((double)_image[index1] - (double)_image[index2]);
 
     double dist = 0.0;
     for (ImgIdx ch = 0; ch < _channels; ch++) {
@@ -23,12 +23,12 @@ template <class Pixel> double PixelDissimilarity<Pixel>::L2(ImgIdx index1, ImgId
         double diff = _image[index1 + offset] - _image[index2 + offset];
         dist += diff * diff;
     }
-    return std::sqrt(dist / (double)_channels);
+    return (float)std::sqrt(dist / (double)_channels);
 }
 
-template <class Pixel> double PixelDissimilarity<Pixel>::LInfinity(ImgIdx index1, ImgIdx index2) const {
+template <class Pixel> float PixelDissimilarity<Pixel>::LInfinity(ImgIdx index1, ImgIdx index2) const {
     if (_channels == 1)
-        return std::abs((double)_image[index1] - (double)_image[index2]);
+        return (float)std::abs((double)_image[index1] - (double)_image[index2]);
 
     double dist = 0.0;
     for (ImgIdx ch = 0; ch < _channels; ch++) {
@@ -36,7 +36,7 @@ template <class Pixel> double PixelDissimilarity<Pixel>::LInfinity(ImgIdx index1
         double diff = std::abs((double)_image[index1 + offset] - (double)_image[index2 + offset]);
         dist = std::max(dist, diff);
     }
-    return dist;
+    return (float)dist;
 }
 
 template class PixelDissimilarity<uint8_t>;
